+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Added
 
+- Added a check to make sure org.clojure/tools.cli is up to date
+
 ## Fixed
 
 ## Changed